% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplotServer}
\alias{barplotServer}
\title{Server Logic for Barplot}
\usage{
barplotServer(id, df, default_fill, default_group)
}
\arguments{
\item{id}{A namespace ID for the module.}

\item{df}{A reactive data frame to be used for plotting.}

\item{default_fill}{Default value for the "Color By" input.}

\item{default_group}{Default value for the "Group By" input.}
}
\value{
A reactive output containing the barplot.
}
\description{
Server Logic for Barplot
}
