% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplotUI}
\alias{barplotUI}
\title{User Interface for Barplot}
\usage{
barplotUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the barplot visualization.
}
\description{
User Interface for Barplot
}
