% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossAssayTab.R
\name{crossAssayTabServer}
\alias{crossAssayTabServer}
\title{Server Logic for Cross-Assay Tab}
\usage{
crossAssayTabServer(id, transcript_data, protein_data, metabolite_data)
}
\arguments{
\item{id}{A namespace ID for the module.}

\item{transcript_data}{Reactive object containing transcriptomic data.}

\item{protein_data}{Reactive object containing proteomic data.}

\item{metabolite_data}{Reactive object containing metabolomic data.}
}
\value{
None; used for rendering UI components and server logic.
}
\description{
Server Logic for Cross-Assay Tab
}
