% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossAssayTab.R
\name{crossAssayTabUI}
\alias{crossAssayTabUI}
\title{User Interface for Cross-Assay Tab}
\usage{
crossAssayTabUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the cross-assay correlation tab.
}
\description{
User Interface for Cross-Assay Tab
}
