% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{download_and_extract}
\alias{download_and_extract}
\title{Download and Extract Data}
\usage{
download_and_extract(url, dest_dir)
}
\arguments{
\item{url}{url for the data located}

\item{dest_dir}{to store the data}
}
\description{
Downloads and extracts the data required for the package.
}
