% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomicTab.R
\name{metabolomicTabServer}
\alias{metabolomicTabServer}
\title{Server Logic for Metabolomic Tab}
\usage{
metabolomicTabServer(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
Reactive data containing metabolomic information.
}
\description{
Server Logic for Metabolomic Tab
}
