% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomicTab.R
\name{metabolomicTabUI}
\alias{metabolomicTabUI}
\title{User Interface for Metabolomic Tab}
\usage{
metabolomicTabUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the metabolomic data tab.
}
\description{
User Interface for Metabolomic Tab
}
