% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkTab.R
\name{networkTabServer}
\alias{networkTabServer}
\title{Server Logic for Network Tab}
\usage{
networkTabServer(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
None; used to manage the network tab's server logic, including node selection, visibility, and zoom.
}
\description{
Server Logic for Network Tab
}
