% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkTab.R
\name{networkTabUI}
\alias{networkTabUI}
\title{User Interface for Network Tab}
\usage{
networkTabUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the network tab, including options for node selection, visibility, and zoom.
}
\description{
User Interface for Network Tab
}
