% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-nodeSelection.R
\name{nodeSelectionServer}
\alias{nodeSelectionServer}
\title{Server Logic for Node Selection}
\usage{
nodeSelectionServer(id, nodes)
}
\arguments{
\item{id}{A namespace ID for the module.}

\item{nodes}{A reactive expression returning a data frame of nodes.}
}
\value{
None; used for managing node selection in the UI.
}
\description{
Server Logic for Node Selection
}
