% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-nodeSelection.R
\name{nodeSelectionUI}
\alias{nodeSelectionUI}
\title{User Interface for Node Selection}
\usage{
nodeSelectionUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the node selection module.
}
\description{
User Interface for Node Selection
}
