% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicTab.R
\name{proteomicTabServer}
\alias{proteomicTabServer}
\title{Server Logic for Proteomic Tab}
\usage{
proteomicTabServer(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
Reactive data containing processed proteomic data.
}
\description{
Server Logic for Proteomic Tab
}
