% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicTab.R
\name{proteomicTabUI}
\alias{proteomicTabUI}
\title{User Interface for Proteomic Tab}
\usage{
proteomicTabUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the proteomic data tab.
}
\description{
User Interface for Proteomic Tab
}
