% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplotServer}
\alias{scatterplotServer}
\title{Server Logic for Scatterplot}
\usage{
scatterplotServer(id, df, default_x, default_y)
}
\arguments{
\item{id}{A namespace ID for the module.}

\item{df}{A reactive data frame to be used for plotting.}

\item{default_x}{Default variable for the X-axis.}

\item{default_y}{Default variable for the Y-axis.}
}
\value{
None; used to render the scatterplot and handle interactions.
}
\description{
Server Logic for Scatterplot
}
