% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplotUI}
\alias{scatterplotUI}
\title{User Interface for Scatterplot}
\usage{
scatterplotUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the scatterplot visualization.
}
\description{
User Interface for Scatterplot
}
