% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTab.R
\name{searchTabServer}
\alias{searchTabServer}
\title{Server Logic for Search Tab}
\usage{
searchTabServer(
  id,
  gene_options,
  protein_options,
  metabolic_options,
  gene_data,
  protein_data_processed,
  metabolic_data_processed
)
}
\arguments{
\item{id}{A namespace ID for the module.}

\item{gene_options}{A reactive list of gene options.}

\item{protein_options}{A reactive list of protein options.}

\item{metabolic_options}{A reactive list of metabolite options.}

\item{gene_data}{A Seurat object containing gene expression data.}

\item{protein_data_processed}{A reactive data frame of processed protein data.}

\item{metabolic_data_processed}{A reactive data frame of processed metabolite data.}
}
\value{
None; used to manage the search tab's server logic.
}
\description{
Server Logic for Search Tab
}
