% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTab.R
\name{searchTabUI}
\alias{searchTabUI}
\title{User Interface for Search Tab}
\usage{
searchTabUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the search tab, allowing users to search and visualize data.
}
\description{
User Interface for Search Tab
}
