% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Launch the Multiomics Explorer Shiny App}
\usage{
server(input, output, session)
}
\arguments{
\item{input, }{input params from UI}

\item{output, }{output params to UI}

\item{session, }{session variable}
}
\value{
the panel for main plots;
}
\description{
This function launches the Multiomics Explorer Shiny application, which provides tools for analyzing and visualizing multi-omics data.
}
