% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApp.R
\name{startApp}
\alias{startApp}
\title{Launch the Multiomics Explorer Shiny App}
\usage{
startApp()
}
\description{
This function launches the Multiomics Explorer Shiny application, which provides tools for analyzing and visualizing multi-omics data.
}
