% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomicTab.R
\name{transcriptomicTabServer}
\alias{transcriptomicTabServer}
\title{Server Logic for Transcriptomic Tab}
\usage{
transcriptomicTabServer(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
Reactive values containing transcriptomic data.
}
\description{
Server Logic for Transcriptomic Tab
}
