% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomicTab.R
\name{transcriptomicTabUI}
\alias{transcriptomicTabUI}
\title{User Interface for Transcriptomic Tab}
\usage{
transcriptomicTabUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for the transcriptomic data tab.
}
\description{
User Interface for Transcriptomic Tab
}
