% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ui}
\alias{ui}
\title{Launch the Multiomics Explorer Shiny App}
\usage{
ui()
}
\description{
This function launches the Multiomics Explorer Shiny application, which provides tools for analyzing and visualizing multi-omics data.
}
