% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-visibility.R
\name{visibilityServer}
\alias{visibilityServer}
\title{Server Logic for Node Visibility Control}
\usage{
visibilityServer(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
None; used to manage node visibility in the UI.
}
\description{
Server Logic for Node Visibility Control
}
