% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-visibility.R
\name{visibilityUI}
\alias{visibilityUI}
\title{User Interface for Node Visibility Control}
\usage{
visibilityUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for controlling the visibility of nodes.
}
\description{
User Interface for Node Visibility Control
}
