% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-zoom.R
\name{zoomServer}
\alias{zoomServer}
\title{Server Logic for Zoom Controls}
\usage{
zoomServer(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
None; used to control zooming of nodes in the UI.
}
\description{
Server Logic for Zoom Controls
}
