% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-zoom.R
\name{zoomUI}
\alias{zoomUI}
\title{User Interface for Zoom Controls}
\usage{
zoomUI(id)
}
\arguments{
\item{id}{A namespace ID for the module.}
}
\value{
A UI definition for zoom controls.
}
\description{
User Interface for Zoom Controls
}
